requirements.txt
streamlit
elasticsearch
sentence-transformers
openai
python-dotenv
tqdm
Pillow
